requirements.txt
flask
flask-mongoengine 
flask-restful
flask-bcrypt 
flask-jwt-extended
flask-cors